Log/FF scales: default to max=1000
Defaulting to 100 seems like a bad idea, because that's the bare minimum for a
"pass" - defaulting to 1000 will hopefully lead to higher expectations (at least
in the context of helping non-experts tests).

OTOH, 100 probably is a reasonable default when testing poor masks for
scientific reasons. Making the default range configurable in future might be a
sensible idea?

Note: these axes are auto-scaling. I.e. the range used to jump to 1k or 10k
as needed (I've never seen anything anywhere near 10k...).