Raw particle graph: implement fixed X axis
(Should this be called a "relative" time axis instead?)

The 8020 will deliver particle readings continuously, even when we're not
actively testing. Viewing this data is potentially useful to verify that
particle counts inside the specimen have stabilised (apparently this can
take a few minutes for some specimens/subjects) - therefore the raw
particle graph will always have data even prior to the test, and therefore
it only makes sense to indicate when we received a sample relative to *now*
as opposed to relative since the time the app started.

(libp8020 does not yet deliver samples until the test has started, but that
 will change soon, hopefully.)

Also switch to using qsizetype (vs size_t) throughout as we're dealing with
negative numbers now.