ExercisesModel: fix renderInterimFF argument count
To match receivedInterimFF signal. This was a regression from
30eeb73f5fc5699be09aebfcb5edf94716430f7d which added an additional parameter to
the signal, without updating corresponding slots. This meant that the slot
always interpreted the exercise number (which is typically a small number) as
the FF.

Surprisingly, Qt does not warn you about this - slots may have fewer parameters
than the signals that they are connected to. This is even more suprising given
that the types aren't matched (although all numeric). Ugghh. Using a
non-implicitly-convertable type might have helped?

Fixes:
https://github.com/ahunt/incolata/issues/19